Add turbolinks for instant page views
Added support for https://github.com/turbolinks/turbolinks, which in my
view is the best what can happen with static site: instead of reloading
pages every time, it turns any static site to single page app, and via
AJAX only fetch changes. It make reloads invisible.

In addition now when you go to the new section, it does not scroll user
to the top of the our nav: instead it scrolls to the top of the
content.

Note that when run `hugo serve` (development mode) by default it inject
live-reload script, which conflicts with Tubmolinks. I turn off
Turbolinks if live-reload script is active, but you can just run `hugo`
with `--disableLiveReload` flag, to test this change on local machine.

In addition I added small fix to docs-table-of-contents.js, which fix
error handling.