Automatic netlify _redirects file generation
Now it is possible to use Hugo Aliases, and it will automatically generate
Netlify _redirects file.

At the moment it is only technical capability, and we still need to ensure that
we migrate existing _redirects file, to use Hugo aliases.

To test:
- Run `hugo`
- It will generate `public` directory, and inside you will find `netlify.json`
  (it is not actually `json` file, it is in netflify format, but just a trick to generate this file)
- Replace _redirects file with netlify.json file

Example of generated _netfliy file:
```
/docs/tyk-developer-portal/keycloak-dcr /docs/nightly/tyk-developer-portal/tyk-portal-classic/keycloak-dcr/ 301
/docs/getting-started/tyk-components/gateway/ /docs/nightly/tyk-oss-gateway/ 301
/docs/concepts/tyk-components/gateway/ /docs/nightly/tyk-oss-gateway/ 301
/docs/graphql/sync-schema/ /docs/nightly/graphql/syncing-schema/ 301
/docs/graphql/introspection/ /docs/nightly/graphql/introspection/ 301
```

As you can see it also adds Metadata in comment, before redirect, about Page
title and its path on the file system. Also `nightly` in redirect target is
fine, since it reads this value from `config.toml` and on stable docs branch it
will have correct value.