Update users.md
Fixed sample response: "User updated" in updated section

Included changes regarding a fix for a customer where they were trying to update their admin account but were being met with an error. This is because the docs did not explicitly mention that they need to include an access_token in the body of the request.