Clarification that this is for HTTP type API (#2839)
Clarification that this is for a HTTP type API only. It can lead to confusion since the Raw API in UDG and GQL also have the possibility of adding persisted_graphql and it gives an error.

(cherry picked from commit aae213d3528e4d0537f336a32a1365f74abb524c)